RAG_XS20_Manual/requirements.txt
langchain
langchain_community
langchain-community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime
